water body segmentation from satellite imagery/requirements.txt
tensorflow==2.17.0
streamlit==1.39.0
opencv-python-headless==4.10.0.84
pillow==10.4.0
numpy==1.26.4